ai-model-call/pyproject.toml
[project]
name = "ai-model-call"
version = "0.1.0"
description = "MCP server for calling various AI models"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = ["fastmcp>=2.0.0", "google-genai>=1.10.0", "openai>=1.72.0"]
